[#8] Cafe JPA Repository 테스트 추가
- CafeRepository 관련 테스트 코드 추가
- p6spy 모듈 적용(테스트 환경만)
- BaseTimeEntity 필드 protected setter 적용
