+++ .gitignore
@@ -34,3 +34,6 @@ out/
 
 ### VS Code ###
 .vscode/
+
+### log ###
+*.log
