+++ src/test/java/io/beaniejoy/dongnecafe/domain/cafe/repository/CafeRepositoryTest.kt
@@ -1,28 +1,41 @@
-package io.beaniejoy.dongnecafe.domain.cafe.entity
+package io.beaniejoy.dongnecafe.domain.cafe.repository
 
 import io.beaniejoy.dongnecafe.domain.cafe.dto.request.CafeInfoRequestDto
 import io.beaniejoy.dongnecafe.domain.cafe.dto.request.CafeMenuInfoRequestDto
 import io.beaniejoy.dongnecafe.domain.cafe.dto.request.MenuOptionInfoRequestDto
 import io.beaniejoy.dongnecafe.domain.cafe.dto.request.OptionDetailInfoRequestDto
+import io.beaniejoy.dongnecafe.domain.cafe.entity.Cafe
 import io.beaniejoy.dongnecafe.domain.cafe.utils.CafeTestUtils
-import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
 import java.math.BigDecimal
 
-internal class CafeTest {
+@DataJpaTest
+internal class CafeRepositoryTest {
+    @Autowired
+    lateinit var cafeRepository: CafeRepository
+
     @Test
-    fun create_cafe_test() {
+    @DisplayName("[JPA] 신규 Cafe 저장 테스트")
+    fun saveTest() {
         val cafeRequestDto = createCreateCafeRequestDto()
+        val cafe = cafeRequestDto.let {
+            Cafe.createCafe(
+                name = it.name!!,
+                address = it.address!!,
+                phoneNumber = it.phoneNumber!!,
+                description = it.description!!,
+                cafeMenuRequestList = it.cafeMenuList
+            )
+        }
 
-        val cafe = Cafe.createCafe(
-            name = cafeRequestDto.name!!,
-            address = cafeRequestDto.address!!,
-            phoneNumber = cafeRequestDto.phoneNumber!!,
-            description = cafeRequestDto.description!!,
-            cafeMenuRequestList = cafeRequestDto.cafeMenuList
-        )
+        val savedCafe = cafeRepository.save(cafe)
 
-        CafeTestUtils.assertCafeEquals(request = cafeRequestDto, entity = cafe)
+        assertEquals(1L, savedCafe.id)
+        CafeTestUtils.assertCafeEquals(cafeRequestDto, savedCafe)
     }
 
     private fun createCreateCafeRequestDto(): CafeInfoRequestDto {
